requirements.txt
accelerate==1.5.2
annotated-types==0.7.0
av==14.2.0
bitsandbytes==0.45.3
blis==1.2.0
Brotli @ file:///croot/brotli-split_1736182456865/work
catalogue==2.0.10
certifi @ file:///croot/certifi_1738623731865/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
decord==0.6.0
easyocr==1.7.2
einops==0.8.1
faiss-gpu==1.7.2
ffmpeg-python==0.2.0
filelock @ file:///croot/filelock_1700591183607/work
flash_attn @ file:///root/nfs/download/flash_attn-2.7.4.post1%2Bcu12torch2.5cxx11abiFALSE-cp310-cp310-linux_x86_64.whl#sha256=b267f80a08e516292cdd748056a2178a45b8abedf7fca123292eb17c21c8c87c
fonttools==4.56.0
fsspec==2025.3.0
future==1.0.0
gmpy2 @ file:///croot/gmpy2_1738085463648/work
huggingface-hub==0.29.3
idna @ file:///croot/idna_1714398848350/work
imageio==2.37.0
Jinja2 @ file:///croot/jinja2_1741710844255/work
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
matplotlib==3.10.1
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///croot/mpmath_1690848262763/work
murmurhash==1.0.12
networkx @ file:///croot/networkx_1737039604450/work
ninja==1.11.1.3
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp310-cp310-linux_x86_64.whl#sha256=d8cd837ed43e87f77e6efaa08e8de927ca030a1c9c5d04624432d6fb9a74a5ee
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow @ file:///croot/pillow_1738010226202/work
preshed==3.0.9
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML @ file:///croot/pyyaml_1728657952215/work
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
rich==13.9.4
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
seaborn==0.13.2
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.13.1
thinc==8.3.4
tifffile==2025.3.13
tokenizers==0.19.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.40.0
triton==3.1.0
typer==0.15.2
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.1
ultralytics==8.3.94
ultralytics-thop==2.0.14
urllib3 @ file:///croot/urllib3_1737133630106/work
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
